+++ docs/CHANGE_LOG.md
@@ -7,6 +7,7 @@
 - The spacing SCSS variable is now a CSS variable (calc() improvements have been included to make size changes easier).
 - The font based SCSS variables are now CSS variables.
 - Removed repetitive code that builds the hours display for the Full Day/Week views.
+- The Full Year view now shows the total number of events that are available per day in each month (little red indicator).
 - 
 
 <br>